Zoom/requirements.txt
opencv-python>=4.7.0
mediapipe>=0.10.11
numpy>=1.23.0
pyvirtualcam>=0.10.0
pystray>=0.19.0
Pillow>=9.0.0
pywin32>=305
